# Improve PropertyMapper performance