### Description of Change

- Improve `PropertyMapper` performance by caching mappers (same thing is currently done on mapper's keys)
  - Moves `viewHandler.CanInvokeMappers()` check to `UpdateProperties` level considering that on Android this method method uses the Java object
- Ensures properties are executed in the order they're defined (this will help with future perf developments #27259)
  - Brings to light race conditions now present on a few mappers (like `MapToolbar` depends on `MapFlyout`) and formally solves them by setting the right order of execution

### Benchmarks

> BenchmarkDotNet v0.13.10, macOS 15.3.1 (24D70) [Darwin 24.3.0]
> Apple M3 Pro, 1 CPU, 11 logical and 11 physical cores
> .NET SDK 9.0.102

**Before**

| Method                    | Mean      | Error    | StdDev   | Gen0      | Allocated  |
|-------------------------- |----------:|---------:|---------:|----------:|-----------:|
| BenchmarkUpdateProperties | 167.67 ms | 0.584 ms | 0.547 ms | 3333.3333 | 29604261 B |
| BenchmarkUpdateProperty   |  31.65 ms | 0.079 ms | 0.070 ms |         - |       46 B |

**After**

| Method                    | Mean     | Error    | StdDev   | Gen0      | Allocated  |
|-------------------------- |---------:|---------:|---------:|----------:|-----------:|
| BenchmarkUpdateProperties | 59.04 ms | 0.294 ms | 0.275 ms | 3666.6667 | 31204122 B |
| BenchmarkUpdateProperty   | 12.39 ms | 0.026 ms | 0.020 ms |         - |       12 B |