@albyrock87 2025-02-28
@mattleibow I've updated the PR, including the fix for the broken tests: I was totally missing the special `AndroidBatchPropertyMapper`.

Also I backported the race-condition fix on iOS `ContainerViewController` I've done in the other PR.