@albyrock87 2025-03-01
Note: these failing tests are unrelated:
- mac: UpdatedIsEnabledProperty.png (size differs - **baseline is 1920x1080 pixels, actual is 789x563 pixels**)
- mac: FlyoutItemTextShouldDisplayProperly.png (size differs - **baseline is 1920x1080 pixels, actual is 789x563 pixels**)
- mac: RemainingItemsThresholdReachedCommandFired => passed locally