@albyrock87 2025-03-03
@mattleibow so I had to refactor the code again due to the `AndroidBatchedPropertyMapper`.

The introduction of that class changed the meaning of `GetKeys`

> returns the list of defined mappers

something like:

> returns the list of keys to be updated in `UpdateProperties`

This is really unfortunate because now there's no way to tell what are the defined keys.

Initially I thought of simply moving the skip list into `PropertyMapper` and restore the original meaning of `GetKeys`, but turns out that I can't because a parent mapper could re-define a given mapper and therefore requiring it's execution.
For example `SwipeViewHandler` redefines `IsEnabled` which is part of the android skip list.

On top of that, if the user ever wanted to define their own `IsEnabled` at `ViewHandler` level (or even `AppendToMapping`), their mapper would never be executed due to the skip list, which is obviously wrong.

#27259 will enable us to detect when were running the initial mapping, so we will be able to remove `AndroidBatchedPropertyMapper` and just turn every skip-list mapper into a noop during the initial mapping, for example:

```
public static void MapVisibility(IViewHandler handler, IView view)
{
#if ANDROID
    if (handler.IsConnectingHandler()) return;
#else
    if (handler.IsConnectingHandler() && view.Visibility == Visibility.Visible) return;
#endif

    if (handler.HasContainer)
        ((PlatformView?)handler.ContainerView)?.UpdateVisibility(view);
```

This way the user can even define their own mappers ensuring they will be executed on the first mapping.