@albyrock87 2025-03-05
Failing tests look unrelated:
- RegisterTwoDoubleTapHandlersAndCheckNumberOfFiredEventsAsync => passes locally
- CarouselViewShouldRenderCorrectly => it's randomly mismatching by 1px (not only in this PR)
- BindingUpdatesFromInteractiveRefresh => I have no clue, and no Windows machine to verify
- CarouselItemsShouldRenderProperly (windows) => the screenshot differs only due to the horizontal scrollbar